requirements.txt
streamlit
tensorflow==2.20.0
numpy
pandas
plotly
pillow